+++ birder/net/rope_vit.py
@@ -871,6 +871,8 @@ class RoPE_ViT(DetectorBackbone, PreTrainEncoder, MaskedTokenOmissionMixin, Mask
         x = self.conv_proj(x)
         x = self.patch_embed(x)
 
+        patch_embedding = x
+
         if self.pos_embed_special_tokens is False:
             x = x + self._get_pos_embed(H, W)
 
@@ -883,14 +885,21 @@ class RoPE_ViT(DetectorBackbone, PreTrainEncoder, MaskedTokenOmissionMixin, Mask
         if len(special_tokens) > 0:
             x = torch.concat(special_tokens + [x], dim=1)
 
+        if return_input_embedding is True:
+            if len(special_tokens) > 0:
+                input_embedding = torch.concat(special_tokens + [patch_embedding], dim=1)
+            else:
+                input_embedding = patch_embedding
+        else:
+            input_embedding = None  # For TorchScript compatibility
+
         if self.pos_embed_special_tokens is True:
             x = x + self._get_pos_embed(H, W)
 
-        input_embedding = x
         x = self.encoder(x, self._get_rope_embed(H, W))
         x = self.norm(x)
 
-        if return_input_embedding is True:
+        if return_input_embedding is True and input_embedding is not None:
             return torch.stack([input_embedding, x], dim=-1)
 
         return x